ml-studio/backend/pyproject.toml
[project]
name = "ml-studio-backend"
version = "0.1.0"
description = "Universal ML comparison backend — FastAPI + MLflow"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "mlflow>=2.13.0",
    "scikit-learn>=1.4.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "requests>=2.31.0",
    "python-multipart>=0.0.9",
    "supabase>=2.4.0",
    "python-dotenv>=1.0.0",
    "optuna>=3.6.0",
    "joblib>=1.4.0",
]

[dependency-groups]
dev = []
